+++ lib/src/appsflyer_sdk.dart
@@ -3,6 +3,7 @@ part of appsflyer_sdk;
 class AppsflyerSdk {
   StreamController _afGCDStreamController;
   StreamController _afOpenAttributionStreamController;
+  StreamController _afValidtaPurchaseController;
   static AppsflyerSdk _instance;
   final MethodChannel _methodChannel;
 
@@ -52,7 +53,7 @@ class AppsflyerSdk {
 
     if (_afGCDStreamController != null) {
       validatedOptions[AppsflyerConstants.AF_GCD] = true;
-      _registerListener();
+      _registerGCDListener();
     } else {
       validatedOptions[AppsflyerConstants.AF_GCD] = false;
     }
@@ -85,7 +86,7 @@ class AppsflyerSdk {
 
     if (_afGCDStreamController != null) {
       afOptions[AppsflyerConstants.AF_GCD] = true;
-      _registerListener();
+      _registerGCDListener();
     } else {
       afOptions[AppsflyerConstants.AF_GCD] = false;
     }
@@ -95,14 +96,34 @@ class AppsflyerSdk {
 
   ///Returns `Stream`. Accessing AppsFlyer Conversion Data from the SDK
   Stream<dynamic> registerConversionDataCallback() {
-    _afGCDStreamController = StreamController();
+    if (_afGCDStreamController == null) {
+      _afGCDStreamController = StreamController(onCancel: () {
+        _afGCDStreamController.close();
+      });
+    }
     return _afGCDStreamController.stream;
   }
 
   ///Returns `Stream`. Accessing AppsFlyer attribution, referred from deep linking
   Stream<dynamic> registerOnAppOpenAttributionCallback() {
-    _afOpenAttributionStreamController = StreamController();
+    if (_afOpenAttributionStreamController == null) {
+      _afOpenAttributionStreamController = StreamController(onCancel: () {
+        _afOpenAttributionStreamController.close();
+      });
+    }
     return _afOpenAttributionStreamController.stream;
+  }
+
+  ///Returns `Stream`. Accessing AppsFlyer purchase validation data
+  Stream<dynamic> _registerValidatePurchaseCallback() {
+    if (_afValidtaPurchaseController == null) {
+      _afValidtaPurchaseController = StreamController(onCancel: () {
+        _afValidtaPurchaseController.close();
+      });
+
+      _registerPurchaseValidateListener();
+    }
+    return _afValidtaPurchaseController.stream;
   }
 
   ///initialize the SDK, using the options initialized from the constructor|
@@ -128,7 +149,112 @@ class AppsflyerSdk {
         "trackEvent", {'eventName': eventName, 'eventValues': eventValues});
   }
 
-  void _registerListener() {
+  void setHost(String hostPrefix, String hostName) {
+    _methodChannel.invokeMethod(
+        "setHost", {'hostPrefix': hostPrefix, 'hostName': hostName});
+  }
+
+  void setCollectIMEI(bool isCollect) {
+    _methodChannel.invokeMethod("setCollectIMEI", {'isCollect': isCollect});
+  }
+
+  void setCollectAndroidId(bool isCollect) {
+    _methodChannel
+        .invokeMethod("setCollectAndroidId", {'isCollect': isCollect});
+  }
+
+  Future<String> getHostName() async {
+    return await _methodChannel.invokeMethod("getHostName");
+  }
+
+  Future<String> getHostPrefix() async {
+    return await _methodChannel.invokeMethod("getHostPrefix");
+  }
+
+  void setAndroidIdData(String androidId) {
+    _methodChannel.invokeMethod("setAndroidIdData", {'androidId': androidId});
+  }
+
+  void setMinTimeBetweenSessions(int seconds) {
+    assert(seconds >= 0, "the minimum timeout must be a positive number");
+    _methodChannel
+        .invokeMethod("setMinTimeBetweenSessions", {'seconds': seconds});
+  }
+
+  void setImeiData(String imei) {
+    _methodChannel.invokeMethod("setImeiData", {'imei': imei});
+  }
+
+  void setCurrencyCode(String currencyCode) {
+    _methodChannel
+        .invokeMethod("setCurrencyCode", {'currencyCode': currencyCode});
+  }
+
+  void setCustomerUserId(String id) {
+    _methodChannel.invokeMethod("setCustomerUserId", {'id': id});
+  }
+
+  void setIsUpdate(bool isUpdate) {
+    _methodChannel.invokeMethod("setIsUpdate", {'isUpdate': isUpdate});
+  }
+
+  void stopTracking(bool isTrackingStopped) {
+    _methodChannel
+        .invokeMethod("stopTracking", {'isTrackingStopped': isTrackingStopped});
+  }
+
+  void enableLocationCollection(bool flag) {
+    _methodChannel.invokeMethod("enableLocationCollection", {'flag': flag});
+  }
+
+  void enableUninstallTracking(String senderId) {
+    _methodChannel
+        .invokeMethod("enableUninstallTracking", {'senderId': senderId});
+  }
+
+  void updateServerUninstallToken(String token) {
+    _methodChannel.invokeMethod("updateServerUninstallToken", {'token': token});
+  }
+
+  void setUserEmails(List<String> emails, [EmailCryptType cryptType]) {
+    if (cryptType != null) {
+      int cryptTypeInt = EmailCryptType.values.indexOf(cryptType);
+      _methodChannel.invokeMethod("setUserEmailsWithCryptType",
+          {'emails': emails, 'cryptType': cryptTypeInt});
+    } else {
+      _methodChannel.invokeMethod("setUserEmails", {'emails': emails});
+    }
+  }
+
+  void waitForCustomerUserId(bool wait) {
+    _methodChannel.invokeMethod("waitForCustomerUserId", {'wait': wait});
+  }
+
+  ///Returns `Stream`. Accessing AppsFlyer purchase validation data
+  Stream<dynamic> validateAndTrackInAppPurchase(
+      String publicKey,
+      String signature,
+      String purchaseData,
+      String price,
+      String currency,
+      Map<String, String> additionalParameters) {
+    _methodChannel.invokeMethod("validateAndTrackInAppPurchase", {
+      'publicKey': publicKey,
+      'signature': signature,
+      'purchaseData': purchaseData,
+      'price': price,
+      'currency': currency,
+      'additionalParameters': additionalParameters
+    });
+    return _registerValidatePurchaseCallback();
+  }
+
+  void setAdditionalData(Map<String, dynamic> customData) {
+    _methodChannel
+        .invokeMethod("setAdditionalData", {'customData': customData});
+  }
+
+  void _registerGCDListener() {
     BinaryMessages.setMessageHandler(AppsflyerConstants.AF_EVENTS_CHANNEL,
         (ByteData message) async {
       final buffer = message.buffer;
@@ -147,6 +273,22 @@ class AppsflyerSdk {
       }
 
       return null;
+    });
+  }
+
+  void _registerPurchaseValidateListener() {
+    BinaryMessages.setMessageHandler(
+        AppsflyerConstants.AF_VALIDATE_PURCHASE_CHANNEL,
+        (ByteData message) async {
+      final buffer = message.buffer;
+      final decodedStr = utf8.decode(buffer.asUint8List());
+      var decodedJSON = jsonDecode(decodedStr);
+      bool success = decodedJSON['success'] == AppsflyerConstants.AF_SUCCESS;
+      if (success) {
+        _afValidtaPurchaseController.sink.add(decodedJSON);
+      } else {
+        _afValidtaPurchaseController.sink.addError(decodedJSON);
+      }
     });
   }
 }